requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1623774798911/work
aiohttp @ file:///home/conda/feedstock_root/build_artifacts/aiohttp_1605734413565/work
argon2-cffi==20.1.0
astor @ file:///home/conda/feedstock_root/build_artifacts/astor_1593610464257/work
astunparse @ file:///home/conda/feedstock_root/build_artifacts/astunparse_1610696312422/work
async-timeout==3.0.1
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1620387926260/work
backcall==0.2.0
bleach==4.0.0
blinker==1.4
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1607575130224/work
brotlipy==0.7.0
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1619593780156/work
certifi==2021.10.8
cffi @ file:///tmp/build/80754af9/cffi_1625814692085/work
chardet==4.0.0
click @ file:///home/conda/feedstock_root/build_artifacts/click_1621503702258/work
coverage @ file:///home/conda/feedstock_root/build_artifacts/coverage_1602253697921/work
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1616851485986/work
cycler==0.10.0
Cython==0.29.22
dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1628958434797/work
debugpy==1.4.1
decorator==4.4.2
defusedxml==0.7.1
dominate==2.6.0
einops==0.3.2
entrypoints==0.3
flatbuffers @ file:///home/conda/feedstock_root/build_artifacts/python-flatbuffers_1620718971772/work
gast @ file:///home/conda/feedstock_root/build_artifacts/gast_1596839682936/work
google-auth @ file:///home/conda/feedstock_root/build_artifacts/google-auth_1629296548061/work
google-auth-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/google-auth-oauthlib_1630497468950/work
google-pasta==0.2.0
grpcio==1.37.0
h5py @ file:///home/conda/feedstock_root/build_artifacts/h5py_1617647517869/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
imageio==2.9.0
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1630267473458/work
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.1
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Preprocessing @ file:///home/conda/feedstock_root/build_artifacts/keras-preprocessing_1610713559828/work
kiwisolver==1.3.1
kmodes==0.11.0
kornia==0.5.2
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1614595805172/work
MarkupSafe==2.0.1
matplotlib==3.4.1
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
multidict @ file:///home/conda/feedstock_root/build_artifacts/multidict_1602413137351/work
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.3
numexpr @ file:///tmp/build/80754af9/numexpr_1618856529730/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1626264704411/work
oauthlib @ file:///home/conda/feedstock_root/build_artifacts/oauthlib_1622563202229/work
opencv-python==4.5.1.48
opt-einsum @ file:///home/conda/feedstock_root/build_artifacts/opt_einsum_1617859230218/work
packaging==21.0
pandas==1.3.4
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@0d53acbcf5123d4c88040fe68fbb9805fc5b2fb9
Pygments==2.10.0
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1619620644068/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1610291451001/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
pyu2f @ file:///home/conda/feedstock_root/build_artifacts/pyu2f_1604248910016/work
PyWavelets==1.1.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.10.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1608156231189/work
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1595492159598/work
rsa @ file:///home/conda/feedstock_root/build_artifacts/rsa_1614171254180/work
scikit-image==0.18.1
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1621370406642/work
scipy @ file:///tmp/build/80754af9/scipy_1618855955483/work
scratchai-nightly==0.0.1a3
Send2Trash==1.8.0
sewar==0.4.4
six @ file:///tmp/build/80754af9/six_1623709665295/work
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-plugin-wit_1611075653546/work/tensorboard_plugin_wit-1.8.0-py3-none-any.whl
tensorflow @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_cpu/tensorflow-base_1614360980945/work/tensorflow_pkg/tensorflow-2.4.1-cp39-cp39-linux_x86_64.whl
tensorflow-estimator @ file:///home/conda/feedstock_root/build_artifacts/tensorflow-estimator_1623773901369/work/wheel_dir/tensorflow_estimator-2.5.0-py2.py3-none-any.whl
termcolor==1.1.0
terminado==0.11.1
testpath==0.5.0
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1628668762525/work
tifffile==2021.4.8
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1622748266870/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1624634538755/work
vit-pytorch==0.20.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1621518206714/work
widgetsnbextension==3.5.1
wrapt @ file:///home/conda/feedstock_root/build_artifacts/wrapt_1606141220149/work
yarl @ file:///home/conda/feedstock_root/build_artifacts/yarl_1605429469302/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1625284368454/work
